pyproject.toml
[project]
name = "top-eleven"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash==2.17.0",
    "dash-bootstrap-components==1.6.0",
    "groq>=0.9.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=0.0.5",
    "langchainhub>=0.1.14",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "plotly==5.22.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "seaborn>=0.13.0",
    "statsmodels>=0.14.0",
    "streamlit>=1.35.0",
    "tavily-python>=0.3.0",
    "xgboost>=2.0.0",
]
